requirements.txt
click
pyyaml
icecream
pydantic
pytest
twitchio
openai